Function ds2::dataflow::parse::create_timely_topology

pub fn create_timely_topology(file: &Path) -> Topology

Parses a Timely dataflow from a Timely log file and returns a topology

Arguments

file: The path to the Timely log file

Comments

The Timely dataflow is extracted from Timely logs of the following form:

Operates(OperatesEvent { id: OperatorId, addr: [WorkerId, X, LocalOperatorId], name: OperatorName }))

Channels(ChannelsEvent { id: ChannelId, scope_addr: [Y, W], source: (LocalOperatorId, Z), target: (LocalOperatorId, P) }))