Function ds2::dataflow::parse::set_source_rates

pub fn set_source_rates(file: &Path, topology: &mut Topology, verbose: bool)

Sets the true output rates of source operators to the values given in a CSV file

Arguments

file: The path to the CSV file

topology: The topology struct to update

verbose: A boolean flag to print further information to standard output

Comments

Each line in file has the form:

source_operator_id,true_output_rate_per_instance