Function ds2::dataflow::parse::update_flink_rates

pub fn update_flink_rates(
    rates_file: &Path, 
    topology: &mut Topology, 
    epoch: Epoch
)

Parses rates of Flink operators from a CSV file and updates a topology

Arguments

rates_file: The path to the CSV file

topology: The topology struct to update with the rate information

epoch: The epoch number the rates correspond to

Comments

Each line in rates_file contains the true and observed rates of an operator instance aggregated per time window. It has the form:

operator_id,instance_id,total_number_of_operator_instances,window_start_timestamp,true_processing_rate,true_output_rate,observed_processing_rate,observed_output_rate