Struct ds2::dataflow::topology::Topology

pub struct Topology {
    pub logical_graph: Graph<Operator, Flow>,
    pub dictionary: HashMap<OperatorId, (NodeIndex, OperatorInstances)>,
    pub channel_dictionary: HashMap<ChannelId, (OperatorId, OperatorId)>,
}

Fields
logical_graph: Graph<Operator, Flow> dictionary: HashMap<OperatorId, (NodeIndex, OperatorInstances)> channel_dictionary: HashMap<ChannelId, (OperatorId, OperatorId)>
Methods

impl Topology

pub fn new() -> Self

pub fn add_unique_node(
&mut self,
id: &str,
name: &str,
instances: u32
) -> NodeIndex

Adds an operator/node to the topology and returns its internal index. If the node exists, it simply returns its index

Arguments

id: The operator id

name: The name of the operator

instances: The total number of operator instances

pub fn add_unique_edge(
&mut self,
source_node: NodeIndex,
target_node: NodeIndex
)

Adds an edge/channel to the topology if it does not exist

Arguments

source_node: The internal index of the source operator/node

target_node: The internal index of the target operator/node

pub fn get_sources_idx(&self) -> HashSet<NodeIndex>

Returns the internal indexes of all source operators/nodes in the topology

pub fn clear_rates(&mut self)

Clears all rate information

pub fn clear(&mut self)

Clears everything

pub fn get_configuration(
&self,
exclude_sources: bool
) -> Vec<(OperatorId, OperatorInstances)>

Retrieves the current dataflow configuration

pub fn set_configuration(&mut self, conf: &Vec<(OperatorId, OperatorInstances)>)

Sets the dataflow configuration

pub fn print(&self)

Prints the topology to standard output