Function ds2::policy::scaling::evaluate_scaling_policy_at_epoch

pub fn evaluate_scaling_policy_at_epoch(
    topo: &mut Topology, 
    threshold: f64, 
    factor: f64, 
    epoch: Epoch, 
    verbose: bool
) -> String

Evaluates the scaling model on a topology with collected metrics for a given epoch

Arguments

Topology: The topology struct with the collected operator rates

threshold: The target rate ratio threshold

factor: The scaling factor

epoch: The epoch for which to evaluate the policy

verbose: A boolean flag to print further information to standard output

Returns

A string of the form:

operator_id1,number_of_instances1,operator_id2,number_of_instances2,...

containing the optimal dataflow configuration, i.e. the optimal level of parallelism for each dataflow operator

Comments

The threshold >= 1 (default: f64::MAX) defines the maximum allowed ratio of a target source rate to the respective observed source rate after the application of a scaling decision. When DS2 converges and the aforementioned ratio for a source operator is above the given threshold, a final configuration adjustment is triggered in an attempt to meet the throughput

The factor >= 0 (default: 0) is used to deliberately over-provision each dataflow operator by factor * 100% more instances than the optimal